Implement SQL saga repository.
This includes changes to the `saga.Repository` interfaces to operate
in terms of "request" and "response" objects. The saga handler has been
updated to suit.

There is an issue with the existing saga correlation mechanism. When
a new message type is added, none of the existing saga instances
will have that message type in their message type.

The solution is probably to allow the application to chose the
keys of the mapping table, rather than using the message type in all cases.