Refactor the saga handler to correct new-instance behaviour of the new mapping system.
The previous change had a design flaw insofar as the onus was put on the mapper
to decide whether a message would be applied to a new saga instance or an
existing instance.

This was possible for keyset based mapper, because it had database entries if
and only the saga instance already exited, but is not possible for the direct
mapper because it is stateless.

This change modifies the interfaces of `saga.Mapper` and `saga.Persister` such
that the mapper is only responsible for deciding whether the message is mapped
to an instance at all, and the persister is responsible for determining whether
the instance already exists or not.

The saga handler calls the not-found handler if the persister returns an
instance at revision zero but the message is not a trigger.