Add `projection` package.
This package defines a `Projector` interface, which is essentially a message
handler that is not able to send messages.

The account projector in the example has been changed to implement this
interface, and an adaptor has been used to allow the projector to be inserted
into the pipeline as a regular message handler.