Refactor the message store in order to implement `persistence.GloballyOrderedStore`.
Part of this refactor involved adding a new interface, `Fetcher`, which hides
the details of fetching messages from the store by stream vs globally.

There's a fair amount of code associated with the message store implementation,
so I've taken this as an opportunity to complete #89 as well. I've mirrored the
structure of the top level Ax packages within the `axmysql` package. Where both
the top-level package and the mysql-specific package are referenced in a single
file, the mysql-specific package is aliased as `mysql<package>`, and the
top-level package is imported without an alias.