Do not obtain a logger from the endpoint.
This was a recent change I made that started us down a path of using the
endpoint as a poor man's service locator, a pattern I'd like to avoid.

This change removes the logger from the `Endpoint` altogether,
instead injecting it directly as a dependency to anything that needs it,
as we should always be doing.

As part of this change I've returned (re)simplified the `observability`
package, removing the ability for observers to access the `Endpoint`.

I suspect long-term that `Endpoint` will go away, replaced with some
convenience functions that wire everything together based on a set of
user-friendly options.